refactor: refactor src/index.t
-src/index.ts: remove the [transform] method and import [src/transform.ts]
-src/transform.ts: add and export the method [transform] remove by index.ts
